java/iot3mobility: remove legacy DENM implementation and fuse EtsiUtils into EtsiConverter + DenmHelper
