+++ java/iot3/mobility/src/main/java/com/orange/iot3mobility/roadobjects/RoadUser.java
@@ -7,7 +7,7 @@
  */
 package com.orange.iot3mobility.roadobjects;
 
-import com.orange.iot3mobility.its.StationType;
+import com.orange.iot3mobility.messages.StationType;
 import com.orange.iot3mobility.messages.cam.core.CamCodec;
 import com.orange.iot3mobility.quadkey.LatLng;
